refactor(core): move authoring inspection into core
