+++ .changeset/early-bikes-tickle.md
@@ -1,0 +1,5 @@
+---
+"skillset": patch
+---
+
+Move authoring inspection and lint orchestration into the core package while keeping CLI compatibility shims.
