+++ packages/core/package.json
@@ -8,6 +8,7 @@
     "./internal/*": "./src/*.ts"
   },
   "dependencies": {
+    "@skillset/lint": "workspace:*",
     "@skillset/transforms": "workspace:*",
     "yaml": "^2.8.1"
   }
